+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 * `Equal` and `EqualCounts`
 * `Size` and `Keys`
 * `IsDisjoint`
+* `SortedUnion`, `SortedIntersect`, and `SortedDifference`
 
 ### Changed
 * `Toggle` values